refactor: 정적페이지를 반환하는 api와 도메인과 관련된 로직을 하는 api 분리
