+++ src/main/java/roomescape/controller/ReservationController.java
@@ -1,32 +1,24 @@
-package roomescape;
+package roomescape.controller;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.Reservation;
+import roomescape.ReservationDto;
 
-@Controller
+@RestController
 public class ReservationController {
 
     private final List<Reservation> reservations = new ArrayList<>();
     private final AtomicLong index = new AtomicLong(1);
-
-    @GetMapping(path = {"/", "/admin"})
-    public String getAdminPage() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String getReservationPage() {
-        return "admin/reservation-legacy";
-    }
 
     @GetMapping("/reservations")
     public ResponseEntity<List<Reservation>> getReservations() {
