+++ src/main/java/roomescape/controller/StaticPageController.java
@@ -1,0 +1,17 @@
+package roomescape.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class StaticPageController {
+    @GetMapping(path = {"/", "/admin"})
+    public String getAdminPage() {
+        return "admin/index";
+    }
+
+    @GetMapping("/admin/reservation")
+    public String getReservationPage() {
+        return "admin/reservation-legacy";
+    }
+}
